Be stricter about ASN.1 decoding
This removes support for ASN.1 indefinite-length encodings, which are
not valid in DER.  It also enforces validity of digits in
GeneralizedTime values, and rejects integers too large to fit in an
intmax_t or uintmax_t.  As a result, four non-exploitable integer
overflows can no longer be triggered.  As a side effect,
interoperability with pre-2000 versions is lost, but those have security
vulnerabilities anyway.

Since left shifts of negative numbers have undefined behavior in C,
replace them with multiplication.

Ideally, the ASN.1 decoder would strictly enforce DER validity.
However, this will require further review of the ecosystem to ensure it
does not cause breakage.